Add "'cobalt_code': 1" to dom.gyp, fix Windows build
To make cobalt_code 1 work, the following changes are made:
- Added MSVC_PUSH_DISABLE_WARNING 4355 to anywhere "this" is used in base
  member initializer list.
- Disabled warning C4702: unreachable code in xtree.
- Added static_cast to all places where implicit conversion will cause warnings,
  or change variable type if it makes sense.
- Fixed all unreferenced parameter warnings.

b/22613037

+++ cobalt/dom/dom_exception.h
@@ -38,7 +38,7 @@ class DOMException : public script::Wrappable {
   explicit DOMException(ExceptionCode code);
   DOMException(ExceptionCode code, const std::string& message);
 
-  uint16 code() const { return code_; }
+  uint16 code() const { return static_cast<uint16>(code_); }
   const std::string& name() const { return name_; }
   const std::string& message() const { return message_; }
